Clean zsh configure files with their integration
Z shell have some configure files. These files are different about a
conditoin shell load them on launch. The conditoin is what kind of call
the shell. For example, login shell, interactive shell or interpreter.
But, it is no need to use each files property in my way to use it. So,
integrate .zshenv and .zprofile into .zshrc.